requirements.txt
asgiref==3.9.1
bleach==6.2.0
dj-database-url==3.0.1
Django==5.2.6
django-csp==4.0
django-heroku==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
packaging==25.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1
whitenoise==6.9.0
